+++ changelog.md
@@ -2,6 +2,7 @@
 ### **Bravas**
 ---
 * LLT-5002: Fix preshared key parsing logic
+* LLT-5004: Wireguard go oscillation
 
 <br>